Adapt UITests for SetupPersonalPinConfirmation Screen, removeduplicated tests
