+++ app/src/androidTestDeviceDebug/kotlin/de/digitalService/useID/SetupIntroTest.kt
@@ -1,18 +1,14 @@
-/*
 package de.digitalService.useID
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import de.digitalService.useID.ui.components.NavigationIcon
 import de.digitalService.useID.ui.screens.setup.SetupIntro
 import de.digitalService.useID.ui.screens.setup.SetupIntroViewModelInterface
-import de.digitalService.useID.util.MockNfcAdapterUtil
-import de.digitalService.useID.util.NfcAdapterUtil
 import de.digitalService.useID.util.setContentUsingUseIdTheme
 import io.mockk.every
 import io.mockk.mockk
@@ -28,9 +24,6 @@ class SetupIntroTest {
 
     @get:Rule(order = 1)
     val composeTestRule = createAndroidComposeRule<MainActivity>()
-
-    @BindValue
-    val mockNfcAdapterUtil: NfcAdapterUtil = MockNfcAdapterUtil()
 
     @Test
     fun shouldShowCancelDialog() {
@@ -103,4 +96,3 @@ class SetupIntroTest {
         verify { viewModel.onNonFirstTimeUsage() }
     }
 }
-*/
