+++ app/src/androidTestDeviceDebug/kotlin/de/digitalService/useID/SetupPersonalPinConfirmTest.kt
@@ -1,17 +1,12 @@
-/*
 package de.digitalService.useID
 
-import androidx.activity.compose.setContent
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
-import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import de.digitalService.useID.ui.screens.setup.SetupPersonalPinConfirm
 import de.digitalService.useID.ui.screens.setup.SetupPersonalPinConfirmViewModel
-import de.digitalService.useID.util.MockNfcAdapterUtil
-import de.digitalService.useID.util.NfcAdapterUtil
+import de.digitalService.useID.ui.screens.setup.SetupPersonalPinConfirmViewModelInterface
 import de.digitalService.useID.util.setContentUsingUseIdTheme
 import io.mockk.every
 import io.mockk.mockk
@@ -28,99 +23,43 @@ class SetupPersonalPinConfirmTest {
     @get:Rule(order = 1)
     val composeTestRule = createAndroidComposeRule<MainActivity>()
 
-    @BindValue
-    val mockNfcAdapterUtil: NfcAdapterUtil = MockNfcAdapterUtil()
-
-    @Test
-    fun inputReceived() {
-        val mockViewModel: SetupPersonalPinConfirmViewModel = mockk(relaxed = true)
-
-        composeTestRule.activity.setContentUsingUseIdTheme {
-            SetupPersonalPinConfirm(viewModel = mockViewModel)
-        }
-
-        composeTestRule.waitForIdle()
-        verify(exactly = 1) { mockViewModel.onInitialize() }
-
-        val pinTitleText = composeTestRule.activity.getString(R.string.firstTimeUser_personalPIN_confirmation_title)
-        composeTestRule.onNodeWithText(pinTitleText).assertIsDisplayed()
-
-        val pinEntryTextFieldTag = "PINEntryField"
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).assertIsFocused()
-
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).performTextInput("1")
-        verify(exactly = 1) { mockViewModel.userInputPin("1") }
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).performTextInput("12")
-        verify(exactly = 1) { mockViewModel.userInputPin("12") }
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).performTextInput("123")
-        verify(exactly = 1) { mockViewModel.userInputPin("123") }
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).performTextInput("1234")
-        verify(exactly = 1) { mockViewModel.userInputPin("1234") }
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).performTextInput("12345")
-        verify(exactly = 1) { mockViewModel.userInputPin("12345") }
-
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).performTextInput("123456")
-        verify(exactly = 1) { mockViewModel.userInputPin("123456") }
-
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).performTextInput("1234567")
-        verify(exactly = 1) { mockViewModel.userInputPin("123456") }
-
-        composeTestRule.onNodeWithTag(pinEntryTextFieldTag).performImeAction()
-        verify(exactly = 1) { mockViewModel.onDoneClicked() }
-    }
+    private val pinEntryFieldTestTag = "PINEntryField"
 
     @Test
     fun correctPinEntryShown() {
-        val testPin = mutableStateOf("")
-
-        val mockViewModel: SetupPersonalPinConfirmViewModel = mockk(relaxed = true)
-        every { mockViewModel.pin } answers { testPin.value }
+        val mockViewModel: SetupPersonalPinConfirmViewModelInterface = mockk(relaxed = true)
+        every { mockViewModel.shouldShowError } returns false
 
         composeTestRule.activity.setContentUsingUseIdTheme {
             SetupPersonalPinConfirm(viewModel = mockViewModel)
         }
 
-        composeTestRule.waitForIdle()
-        verify(exactly = 1) { mockViewModel.onInitialize() }
-
-        val transportPinDialogTitleText = composeTestRule.activity.getString(R.string.firstTimeUser_personalPIN_confirmation_title)
-        composeTestRule.onNodeWithText(transportPinDialogTitleText).assertIsDisplayed()
+        val personalPinConfirmationTitleText = composeTestRule.activity.getString(R.string.firstTimeUser_personalPIN_confirmation_title)
+        composeTestRule.onNodeWithText(personalPinConfirmationTitleText).assertIsDisplayed()
 
         val pinEntryTestTag = "Obfuscation"
         composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(0)
 
-        testPin.value = "1"
+        composeTestRule.onNodeWithTag(pinEntryFieldTestTag).performTextInput("1")
         composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(1)
 
-        testPin.value = "12"
+        composeTestRule.onNodeWithTag(pinEntryFieldTestTag).performTextInput("2")
         composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(2)
 
-        testPin.value = "123"
+        composeTestRule.onNodeWithTag(pinEntryFieldTestTag).performTextInput("3")
         composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(3)
 
-        testPin.value = "1234"
+        composeTestRule.onNodeWithTag(pinEntryFieldTestTag).performTextInput("4")
         composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(4)
 
-        testPin.value = "12345"
+        composeTestRule.onNodeWithTag(pinEntryFieldTestTag).performTextInput("5")
         composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(5)
 
-        testPin.value = "1"
-        composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(1)
-
-        testPin.value = "12345"
-        composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(5)
-
-        testPin.value = "123"
-        composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(3)
-
-        testPin.value = ""
-        composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(0)
-
-        testPin.value = "1234567890"
+        composeTestRule.onNodeWithTag(pinEntryFieldTestTag).performTextInput("6")
         composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(6)
 
-        testPin.value = "1234"
-        composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(4)
+        composeTestRule.onNodeWithTag(pinEntryFieldTestTag).performTextInput("7")
+        composeTestRule.onAllNodesWithTag(pinEntryTestTag).assertCountEquals(6)
     }
 
     @Test
@@ -133,9 +72,6 @@ class SetupPersonalPinConfirmTest {
             SetupPersonalPinConfirm(viewModel = mockViewModel)
         }
 
-        composeTestRule.waitForIdle()
-        verify(exactly = 1) { mockViewModel.onInitialize() }
-
         val errorDialogTitleText = composeTestRule.activity.getString(R.string.firstTimeUser_personalPIN_error_mismatch_title)
         composeTestRule.onNodeWithText(errorDialogTitleText).assertIsDisplayed()
 
@@ -145,4 +81,3 @@ class SetupPersonalPinConfirmTest {
         verify(exactly = 1) { mockViewModel.onErrorDialogButtonClicked() }
     }
 }
-*/
